+++ widgets/src/main/java/com/cassidy/widgets/text/amount/models/Amount.kt
@@ -1,7 +1,9 @@
 package com.cassidy.widgets.text.amount.models
 
+import java.math.BigDecimal
+
 /**
  * @author Gonzalo Cantarero Pérez
  */
 
-data class Amount(val value: Double, val code: String? = null)
+data class Amount(val value: BigDecimal, val code: String? = null)
